feat: refactor context handling for repositories
