+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
       - id: trailing-whitespace
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.12.5
+    rev: v0.12.7
     hooks:
       - id: ruff-check
         args:
@@ -28,3 +28,4 @@ repos:
       entry: uv run mypy
       language: system
       types: [python]
+      exclude: ^.venv/
