+++ app/api/dependencies.py
@@ -8,7 +8,7 @@ from fastapi.security import OAuth2PasswordBearer
 from jwt import InvalidTokenError
 
 from app.core.config import Settings
-from app.core.context.sql import SqlContext
+from app.core.context.context import Context
 from app.domain.context import ContextProtocol
 from app.domain.domain import Domain
 from app.domain.users.entities import User
@@ -29,7 +29,7 @@ def get_settings() -> Settings:
 def get_context(
     settings: Annotated[Settings, Depends(get_settings)],
 ) -> ContextProtocol:
-    return SqlContext(settings=settings)
+    return Context(settings=settings)
 
 
 def get_domain(
