+++ app/cli/main.py
@@ -4,11 +4,11 @@ from pathlib import Path
 
 from app.cli.app import create_cli_app
 from app.core.config import Settings
-from app.core.context.sql import SqlContext
+from app.core.context.context import Context
 from app.domain.domain import Domain
 
 settings = Settings(_env_file=".env")
-context = SqlContext(settings=settings)
+context = Context(settings=settings)
 domain = Domain(context=context)
