+++ app/core/context/sql.py
@@ -9,12 +9,6 @@ from sqlalchemy.orm import Session, sessionmaker
 
 from app.core.config import Settings
 from app.domain.context import ContextProtocol
-from app.domain.interfaces.task_queue import TaskQueueProtocol
-from app.domain.posts.repository import PostRepositoryProtocol
-from app.domain.users.repository import UserRepositoryProtocol
-from app.infrastructure.celery_task_queue.celery_task_queue import CeleryTaskQueue
-from app.infrastructure.sql.posts import PostSqlRepository
-from app.infrastructure.sql.users import UserSqlRepository
 
 logger = logging.getLogger(__name__)
 
@@ -62,15 +56,3 @@ class SqlContext(ContextProtocol):
         if self._session is None:
             raise RuntimeError("No active session.")
         return self._session
-
-    @property
-    def post_repository(self) -> PostRepositoryProtocol:
-        return PostSqlRepository(session=self.session)
-
-    @property
-    def user_repository(self) -> UserRepositoryProtocol:
-        return UserSqlRepository(session=self.session)
-
-    @property
-    def task_queue(self) -> TaskQueueProtocol:
-        return CeleryTaskQueue(settings=self.settings)
