+++ app/domain/context.py
@@ -1,13 +1,13 @@
 from typing import Protocol
 
-from cleanstack.domain import BaseContextProtocol
+from cleanstack.domain import UnitOfWorkProtocol
 
 from app.domain.interfaces.task_queue import TaskQueueProtocol
 from app.domain.posts.repository import PostRepositoryProtocol
 from app.domain.users.repository import UserRepositoryProtocol
 
 
-class ContextProtocol(BaseContextProtocol, Protocol):
+class ContextProtocol(UnitOfWorkProtocol, Protocol):
     @property
     def post_repository(self) -> PostRepositoryProtocol: ...
     @property
